pyproject.toml
[project]
name = "kai-erp-connector"
version = "3.0.0"
description = "KAI ERP Connector - Making ERP Data AI-Ready"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [
    { name = "Kaspar Companies IT", email = "it@kasparcompanies.com" }
]

dependencies = [
    # Core
    "httpx>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Data processing
    "duckdb>=0.10.0",
    "pandas>=2.1.0",
    
    # API
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "slowapi>=0.1.9",  # Rate limiting
    "python-jose[cryptography]>=3.3.0",  # JWT support
    "prometheus-client>=0.19.0",  # Metrics
    
    # MCP (Model Context Protocol)
    "mcp>=1.0.0",
    
    # Async
    "anyio>=4.2.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "respx>=0.20.0",  # HTTP mocking for httpx
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "pandas-stubs>=2.1.0",
    "pre-commit>=3.5.0",
    "bandit[toml]>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/kai_erp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.bandit]
exclude_dirs = ["tests", "docs"]
skips = ["B101"]  # Skip assert_used check (used in tests)

[project.scripts]
kai-erp = "kai_erp.api.main:run"
kai-erp-mcp = "kai_erp.mcp.server:run"
